Rumors System
Self-deteriorating and deleting Rumors System 

+rumors OR +rumours - Displays all categories of rumors and their total in each.
+rumors <category>    - Displays all rumors within that particular category.
+rumors/add <category>=<text> - Adds a new rumor within that particular category.
+rumors/del <category>/<number> - Deletes a certain rumor- limited to those who return 1 for CAN or the person who started the rumor.

CAT is just an attribute tree branch.
CAT`<category> is the branch for a category.
CAT`<category>`<number> is an individual rumor, sorted by number.
CAT`<categpry>`<number>`READ lists dbrefs of all players who have read that specific rumor.
CAT`<category>`<number>`BY is the person who started the rumor.

Whenever a rumor is read within its category, each rumor is checked to see if the enactor has read each rumor. If so, nothing more will happen. If not, the enactor is added to the *`BY attribute and the rumor is checked to see if 1) It's a third read, meaning a random non-ellipsed word gets replaced with ellipses 2) It gets purged due to being composed of 60% ellipses.

+rumors/newcategory and /deletecategory should be self explanatory- only those who pass CAN are able to do this.